api/requirements.txt
openapi-schema-pydantic==1.2.4
pydantic==1.10.9
requests==2.31.0
gunicorn==19.9.0
openai
google.auth
google-cloud-aiplatform
tiktoken
fastapi
uvicorn
pika
celery
redis
uuid
boto3